In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	unsigned int a = 1;
-	unsigned int b = 2;
+	unsigned long int a = 1;
+	unsigned long int b = 2;
 	int x;
-	unsigned int res;
+	unsigned long int res;
 
 	for (x = 1; x <= 99; x++)
 	{
-		printf("%u", a);
+		printf("%lu", a);
 		if (x == 99)
 			break;
 		putchar(',');
